Adding ability to exclude records from storage to reduce unrequired database write -- Fixes #46

+++ src/record/record-deletion.js
@@ -22,10 +22,14 @@ var RecordDeletion = function( options, socketWrapper, message, successCallback 
 	this._isDestroyed = false;
 
 	this._cacheTimeout = setTimeout( this._handleError.bind( this, 'cache timeout' ), this._options.cacheRetrievalTimeout );
-	this._storageTimeout = setTimeout( this._handleError.bind( this, 'storage timeout' ), this._options.storageRetrievalTimeout );
-
 	this._options.cache.delete( this._recordName, this._checkIfDone.bind( this, this._cacheTimeout ) );
-	this._options.storage.delete(this._recordName, this._checkIfDone.bind( this, this._storageTimeout ) );
+
+	if( !this._options.storageExclusion || !this._options.storageExclusion.test( this._recordName ) ) {
+		this._storageTimeout = setTimeout( this._handleError.bind( this, 'storage timeout' ), this._options.storageRetrievalTimeout );
+		this._options.storage.delete(this._recordName, this._checkIfDone.bind( this, this._storageTimeout ) );
+	} else {
+		this._checkIfDone( null );
+	}
 };
 
 /**
